Added Compiled Binaries
This exporter is functional. The options panel doesn't work because I
don't know how to get it to work. Its also designed around quirks of the
Irrlicht DirectX Model Importer and so won't work with straight DirectX
(Somewhere I have a version that works with DirectX and Irrlicht) But
here it is bugs and all for Maya 2011